refactor(broker): extract creation of zeebe partitions
