+++ broker/src/main/java/io/camunda/zeebe/broker/partitioning/PartitionManagerImpl.java
@@ -7,69 +7,32 @@
  */
 package io.camunda.zeebe.broker.partitioning;
 
-import io.atomix.cluster.ClusterMembershipService;
-import io.atomix.cluster.MemberId;
-import io.atomix.cluster.messaging.ClusterCommunicationService;
-import io.atomix.cluster.messaging.ClusterEventService;
 import io.atomix.primitive.partition.ManagedPartitionGroup;
 import io.atomix.primitive.partition.ManagedPartitionService;
 import io.atomix.primitive.partition.impl.DefaultPartitionService;
-import io.atomix.raft.partition.RaftPartition;
 import io.atomix.raft.partition.RaftPartitionGroup;
 import io.atomix.utils.concurrent.Futures;
 import io.camunda.zeebe.broker.PartitionListener;
 import io.camunda.zeebe.broker.clustering.ClusterServices;
-import io.camunda.zeebe.broker.engine.impl.DeploymentDistributorImpl;
-import io.camunda.zeebe.broker.engine.impl.LongPollingJobNotification;
-import io.camunda.zeebe.broker.engine.impl.PartitionCommandSenderImpl;
 import io.camunda.zeebe.broker.exporter.repo.ExporterRepository;
 import io.camunda.zeebe.broker.partitioning.topology.TopologyManager;
 import io.camunda.zeebe.broker.partitioning.topology.TopologyManagerImpl;
 import io.camunda.zeebe.broker.partitioning.topology.TopologyPartitionListener;
-import io.camunda.zeebe.broker.partitioning.topology.TopologyPartitionListenerImpl;
 import io.camunda.zeebe.broker.system.configuration.BrokerCfg;
 import io.camunda.zeebe.broker.system.management.deployment.PushDeploymentRequestHandler;
 import io.camunda.zeebe.broker.system.monitoring.BrokerHealthCheckService;
 import io.camunda.zeebe.broker.system.monitoring.DiskSpaceUsageListener;
-import io.camunda.zeebe.broker.system.partitions.PartitionBoostrapAndTransitionContextImpl;
-import io.camunda.zeebe.broker.system.partitions.PartitionBootstrapStep;
 import io.camunda.zeebe.broker.system.partitions.PartitionHealthBroadcaster;
-import io.camunda.zeebe.broker.system.partitions.PartitionStep;
-import io.camunda.zeebe.broker.system.partitions.PartitionStepMigrationHelper;
-import io.camunda.zeebe.broker.system.partitions.PartitionTransitionStep;
-import io.camunda.zeebe.broker.system.partitions.TypedRecordProcessorsFactory;
 import io.camunda.zeebe.broker.system.partitions.ZeebePartition;
-import io.camunda.zeebe.broker.system.partitions.impl.AtomixPartitionMessagingService;
-import io.camunda.zeebe.broker.system.partitions.impl.PartitionProcessingState;
-import io.camunda.zeebe.broker.system.partitions.impl.PartitionTransitionImpl;
-import io.camunda.zeebe.broker.system.partitions.impl.steps.ExporterDirectorPartitionStep;
-import io.camunda.zeebe.broker.system.partitions.impl.steps.LogDeletionPartitionStep;
-import io.camunda.zeebe.broker.system.partitions.impl.steps.LogStreamPartitionStep;
-import io.camunda.zeebe.broker.system.partitions.impl.steps.RocksDbMetricExporterPartitionStep;
-import io.camunda.zeebe.broker.system.partitions.impl.steps.SnapshotDirectorPartitionStep;
-import io.camunda.zeebe.broker.system.partitions.impl.steps.SnapshotReplicationPartitionBootstrapStep;
-import io.camunda.zeebe.broker.system.partitions.impl.steps.StateControllerPartitionStep;
-import io.camunda.zeebe.broker.system.partitions.impl.steps.StoragePartitionTransitionStep;
-import io.camunda.zeebe.broker.system.partitions.impl.steps.StreamProcessorPartitionStep;
-import io.camunda.zeebe.broker.system.partitions.impl.steps.ZeebeDbPartitionStep;
 import io.camunda.zeebe.broker.transport.commandapi.CommandApiService;
-import io.camunda.zeebe.engine.processing.EngineProcessors;
-import io.camunda.zeebe.engine.processing.message.command.SubscriptionCommandSender;
-import io.camunda.zeebe.engine.processing.streamprocessor.ProcessingContext;
-import io.camunda.zeebe.engine.processing.streamprocessor.TypedRecord;
-import io.camunda.zeebe.engine.processing.streamprocessor.writers.CommandResponseWriter;
-import io.camunda.zeebe.engine.state.mutable.MutableZeebeState;
-import io.camunda.zeebe.logstreams.log.LogStream;
 import io.camunda.zeebe.protocol.impl.encoding.BrokerInfo;
 import io.camunda.zeebe.snapshots.impl.FileBasedSnapshotStoreFactory;
 import io.camunda.zeebe.util.health.HealthStatus;
-import io.camunda.zeebe.util.sched.ActorControl;
 import io.camunda.zeebe.util.sched.ActorSchedulingService;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
-import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -80,37 +43,6 @@ public final class PartitionManagerImpl implements PartitionManager, TopologyMan
 
   private static final Logger LOGGER =
       LoggerFactory.getLogger("io.camunda.zeebe.broker.partitioning");
-
-  // preparation for future steps
-  // will be executed in the order they are defined in this list
-  private static final List<PartitionBootstrapStep> BOOTSTRAP_STEPS =
-      List.of(new SnapshotReplicationPartitionBootstrapStep());
-
-  // will probably be executed in parallel
-  private static final List<PartitionTransitionStep> TRANSITION_STEPS =
-      List.of(new StoragePartitionTransitionStep());
-  // preparation for future step
-
-  private static final List<PartitionStep> LEADER_STEPS =
-      List.of(
-          PartitionStepMigrationHelper.fromBootstrapStep(
-              new SnapshotReplicationPartitionBootstrapStep()),
-          new StateControllerPartitionStep(),
-          PartitionStepMigrationHelper.fromTransitionStep(new StoragePartitionTransitionStep()),
-          new LogDeletionPartitionStep(),
-          new LogStreamPartitionStep(),
-          new ZeebeDbPartitionStep(),
-          new StreamProcessorPartitionStep(),
-          new SnapshotDirectorPartitionStep(),
-          new RocksDbMetricExporterPartitionStep(),
-          new ExporterDirectorPartitionStep());
-  private static final List<PartitionStep> FOLLOWER_STEPS =
-      List.of(
-          PartitionStepMigrationHelper.fromBootstrapStep(
-              new SnapshotReplicationPartitionBootstrapStep()),
-          new StateControllerPartitionStep(),
-          PartitionStepMigrationHelper.fromTransitionStep(new StoragePartitionTransitionStep()),
-          new LogDeletionPartitionStep());
 
   protected volatile CompletableFuture<Void> closeFuture;
   private final BrokerHealthCheckService healthCheckService;
@@ -160,121 +92,14 @@ public final class PartitionManagerImpl implements PartitionManager, TopologyMan
 
     final var membershipService = clusterServices.getMembershipService();
     final var communicationService = clusterServices.getCommunicationService();
-    final var eventService = clusterServices.getEventService();
 
-    partitionService = buildPartitionService(membershipService, communicationService);
+    partitionService =
+        new DefaultPartitionService(membershipService, communicationService, partitionGroup);
 
     topologyManager = new TopologyManagerImpl(membershipService, localBroker);
     partitionListeners.add(topologyManager);
 
     this.partitionListeners = partitionListeners;
-  }
-
-  private void constructPartitions(
-      final ActorSchedulingService actorSchedulingService,
-      final BrokerCfg brokerCfg,
-      final BrokerInfo localBroker,
-      final PushDeploymentRequestHandler deploymentRequestHandler,
-      final List<PartitionListener> partitionListeners,
-      final CommandApiService commandHApiService,
-      final FileBasedSnapshotStoreFactory snapshotStoreFactory,
-      final RaftPartitionGroup partitionGroup,
-      final ClusterMembershipService membershipService,
-      final ClusterCommunicationService communicationService,
-      final ClusterEventService eventService) {
-    final MemberId nodeId = membershipService.getLocalMember().id();
-
-    final List<RaftPartition> owningPartitions =
-        partitionGroup.getPartitionsWithMember(nodeId).stream()
-            .map(RaftPartition.class::cast)
-            .collect(Collectors.toList());
-
-    final TypedRecordProcessorsFactory typedRecordProcessorsFactory =
-        buildTypedRecordProcessorsFactory(
-            localBroker, deploymentRequestHandler, communicationService, eventService);
-
-    for (final RaftPartition owningPartition : owningPartitions) {
-      final var partitionId = owningPartition.id().id();
-
-      final PartitionBoostrapAndTransitionContextImpl transitionContext =
-          buildTransitionContext(
-              actorSchedulingService,
-              brokerCfg,
-              localBroker,
-              partitionListeners,
-              () -> commandHApiService.newCommandResponseWriter(),
-              () -> commandHApiService.getOnProcessedListener(partitionId),
-              membershipService,
-              communicationService,
-              snapshotStoreFactory,
-              typedRecordProcessorsFactory,
-              exporterRepository,
-              owningPartition,
-              partitionId);
-      final PartitionTransitionImpl transitionBehavior =
-          new PartitionTransitionImpl(transitionContext, LEADER_STEPS, FOLLOWER_STEPS);
-      final ZeebePartition zeebePartition =
-          new ZeebePartition(transitionContext, transitionBehavior);
-      partitions.add(zeebePartition);
-    }
-  }
-
-  private PartitionBoostrapAndTransitionContextImpl buildTransitionContext(
-      final ActorSchedulingService actorSchedulingService,
-      final BrokerCfg brokerCfg,
-      final BrokerInfo localBroker,
-      final List<PartitionListener> partitionListeners,
-      final Supplier<CommandResponseWriter> commandResponseWriterSupplier,
-      final Supplier<Consumer<TypedRecord<?>>> onProcessedListenerSupplier,
-      final ClusterMembershipService membershipService,
-      final ClusterCommunicationService communicationService,
-      final FileBasedSnapshotStoreFactory snapshotStoreFactory,
-      final TypedRecordProcessorsFactory typedRecordProcessorsFactory,
-      final ExporterRepository exporterRepository,
-      final RaftPartition owningPartition,
-      final Integer partitionId) {
-    final PartitionBoostrapAndTransitionContextImpl transitionContext =
-        new PartitionBoostrapAndTransitionContextImpl(
-            localBroker.getNodeId(),
-            owningPartition,
-            partitionListeners,
-            buildPartitionMessagingService(
-                membershipService, communicationService, owningPartition),
-            actorSchedulingService,
-            brokerCfg,
-            commandResponseWriterSupplier,
-            onProcessedListenerSupplier,
-            snapshotStoreFactory.getConstructableSnapshotStore(partitionId),
-            snapshotStoreFactory.getReceivableSnapshotStore(partitionId),
-            typedRecordProcessorsFactory,
-            exporterRepository,
-            new PartitionProcessingState(owningPartition));
-    return transitionContext;
-  }
-
-  private TypedRecordProcessorsFactory buildTypedRecordProcessorsFactory(
-      final BrokerInfo localBroker,
-      final PushDeploymentRequestHandler deploymentRequestHandler,
-      final ClusterCommunicationService communicationService,
-      final ClusterEventService eventService) {
-    final var typedRecordProcessorsFactory =
-        createFactory(
-            this::addTopologyPartitionListener,
-            localBroker,
-            communicationService,
-            eventService,
-            deploymentRequestHandler);
-    return typedRecordProcessorsFactory;
-  }
-
-  private AtomixPartitionMessagingService buildPartitionMessagingService(
-      final ClusterMembershipService membershipService,
-      final ClusterCommunicationService communicationService,
-      final RaftPartition owningPartition) {
-    final var messagingService =
-        new AtomixPartitionMessagingService(
-            communicationService, membershipService, owningPartition.members());
-    return messagingService;
   }
 
   @Override
@@ -297,22 +122,20 @@ public final class PartitionManagerImpl implements PartitionManager, TopologyMan
             ps -> {
               LOGGER.info("Starting partitions");
 
-              final var membershipService = clusterServices.getMembershipService();
-              final var communicationService = clusterServices.getCommunicationService();
-              final var eventService = clusterServices.getEventService();
+              final var partitionFactory =
+                  new PartitionFactory(
+                      actorSchedulingService,
+                      brokerCfg,
+                      localBroker,
+                      deploymentRequestHandler,
+                      commandHApiService,
+                      snapshotStoreFactory,
+                      clusterServices,
+                      exporterRepository);
 
-              constructPartitions(
-                  actorSchedulingService,
-                  brokerCfg,
-                  localBroker,
-                  deploymentRequestHandler,
-                  partitionListeners,
-                  commandHApiService,
-                  snapshotStoreFactory,
-                  partitionGroup,
-                  membershipService,
-                  communicationService,
-                  eventService);
+              partitions.addAll(
+                  partitionFactory.constructPartitions(
+                      partitionGroup, partitionListeners, this::addTopologyPartitionListener));
 
               final var futures =
                   partitions.stream()
@@ -375,17 +198,16 @@ public final class PartitionManagerImpl implements PartitionManager, TopologyMan
     partition.close();
   }
 
-  /** Builds a partition service. */
-  private ManagedPartitionService buildPartitionService(
-      final ClusterMembershipService clusterMembershipService,
-      final ClusterCommunicationService messagingService) {
-
-    return new DefaultPartitionService(clusterMembershipService, messagingService, partitionGroup);
-  }
-
   @Override
   public String toString() {
-    return "PartitionManagerImpl{" + "partitionGroup=" + partitionGroup + '}';
+    return "PartitionManagerImpl{"
+        + "partitionService="
+        + partitionService
+        + ", partitionGroup="
+        + partitionGroup
+        + ", partitions="
+        + partitions
+        + '}';
   }
 
   public void onHealthChanged(final int i, final HealthStatus healthStatus) {
@@ -400,43 +222,6 @@ public final class PartitionManagerImpl implements PartitionManager, TopologyMan
   @Override
   public void addTopologyPartitionListener(final TopologyPartitionListener listener) {
     topologyManager.addTopologyPartitionListener(listener);
-  }
-
-  private TypedRecordProcessorsFactory createFactory(
-      final Consumer<TopologyPartitionListener> partitionListenerConsumer,
-      final BrokerInfo localBroker,
-      final ClusterCommunicationService communicationService,
-      final ClusterEventService eventService,
-      final PushDeploymentRequestHandler deploymentRequestHandler) {
-    return (ActorControl actor,
-        MutableZeebeState zeebeState,
-        ProcessingContext processingContext) -> {
-      final LogStream stream = processingContext.getLogStream();
-
-      final TopologyPartitionListenerImpl partitionListener =
-          new TopologyPartitionListenerImpl(actor);
-      partitionListenerConsumer.accept(partitionListener);
-
-      final DeploymentDistributorImpl deploymentDistributor =
-          new DeploymentDistributorImpl(
-              communicationService, eventService, partitionListener, actor);
-
-      final PartitionCommandSenderImpl partitionCommandSender =
-          new PartitionCommandSenderImpl(communicationService, partitionListenerConsumer, actor);
-      final SubscriptionCommandSender subscriptionCommandSender =
-          new SubscriptionCommandSender(stream.getPartitionId(), partitionCommandSender);
-
-      final LongPollingJobNotification jobsAvailableNotification =
-          new LongPollingJobNotification(eventService);
-
-      return EngineProcessors.createEngineProcessors(
-          processingContext,
-          localBroker.getPartitionsCount(),
-          subscriptionCommandSender,
-          deploymentDistributor,
-          deploymentRequestHandler,
-          jobsAvailableNotification::onJobsAvailable);
-    };
   }
 
   public List<ZeebePartition> getPartitions() {
